# Implement property dependency injection for the website