--- a/ProjectLighthouse.Servers.API/Program.cs
+++ b/ProjectLighthouse.Servers.API/Program.cs
@@ -1,36 +1,22 @@
+using LBPUnion.ProjectLighthouse;
 using LBPUnion.ProjectLighthouse.Configuration;
 using LBPUnion.ProjectLighthouse.Logging.Loggers.AspNet;
 using LBPUnion.ProjectLighthouse.Servers.API.Startup;
 using LBPUnion.ProjectLighthouse.Types.Misc;
-using Microsoft.Extensions.DependencyInjection.Extensions;
 
-namespace LBPUnion.ProjectLighthouse.Servers.API;
+await StartupTasks.Run(ServerType.Api);
 
-public static class Program
+IHostBuilder builder = Host.CreateDefaultBuilder();
+builder.ConfigureWebHostDefaults(webBuilder =>
 {
-    public static void Main(string[] args)
-    {
-        StartupTasks.Run(args, ServerType.Api);
+    webBuilder.UseStartup<ApiStartup>();
+    webBuilder.UseUrls(ServerConfiguration.Instance.ApiListenUrl);
+});
 
-        CreateHostBuilder(args).Build().Run();
-    }
+builder.ConfigureLogging(logging =>
+{
+    logging.ClearProviders();
+    logging.AddProvider(new AspNetToLighthouseLoggerProvider());
+});
 
-    public static IHostBuilder CreateHostBuilder(string[] args)
-        => Host.CreateDefaultBuilder(args)
-            .ConfigureWebHostDefaults
-            (
-                webBuilder =>
-                {
-                    webBuilder.UseStartup<ApiStartup>();
-                    webBuilder.UseUrls(ServerConfiguration.Instance.ApiListenUrl);
-                }
-            )
-            .ConfigureLogging
-            (
-                logging =>
-                {
-                    logging.ClearProviders();
-                    logging.Services.TryAddEnumerable(ServiceDescriptor.Singleton<ILoggerProvider, AspNetToLighthouseLoggerProvider>());
-                }
-            );
-}
+await builder.Build().RunAsync();