--- a/ProjectLighthouse.Servers.API/Responses/ApiSlot.cs
+++ b/ProjectLighthouse.Servers.API/Responses/ApiSlot.cs
@@ -68,7 +68,7 @@ public struct ApiSlot
             PlaysUnique = slot.PlaysUnique,
             PlaysComplete = slot.PlaysComplete,
             CommentsEnabled = slot.CommentsEnabled,
-            AverageRating = slot.RatingLBP1,
+            AverageRating = slot.RatingLBP1(context),
             LevelType = slot.LevelType,
         };
 }