--- a/ProjectLighthouse.Servers.Website/Pages/Layouts/BaseLayout.cshtml.cs
+++ b/ProjectLighthouse.Servers.Website/Pages/Layouts/BaseLayout.cshtml.cs
@@ -29,7 +29,7 @@ public class BaseLayout : PageModel
     public BaseLayout(DatabaseContext database)
     {
         this.Database = database;
-        
+
         this.NavigationItems.Add(new PageNavigationItem(BaseLayoutStrings.HeaderUsers, "/users/0", "user friends"));
         this.NavigationItems.Add(new PageNavigationItem(BaseLayoutStrings.HeaderPhotos, "/photos/0", "camera"));
         this.NavigationItems.Add(new PageNavigationItem(BaseLayoutStrings.HeaderSlots, "/slots/0", "globe americas"));