--- a/ProjectLighthouse.Servers.Website/Pages/Partials/CommentsPartial.cshtml
+++ b/ProjectLighthouse.Servers.Website/Pages/Partials/CommentsPartial.cshtml
@@ -1,9 +1,11 @@
 ﻿@using System.Web
 @using System.IO
+@using LBPUnion.ProjectLighthouse.Database
 @using LBPUnion.ProjectLighthouse.Localization
 @using LBPUnion.ProjectLighthouse.Servers.Website.Extensions
 @using LBPUnion.ProjectLighthouse.Types.Entities.Interaction
 @using LBPUnion.ProjectLighthouse.Types.Entities.Profile
+@inject DatabaseContext Database
 
 @{
     string language = (string?)ViewData["Language"] ?? LocalizationManager.DefaultLang;
@@ -50,7 +52,7 @@
             int yourThumb = commentAndReaction.Value?.Rating ?? 0;
             DateTimeOffset timestamp = DateTimeOffset.FromUnixTimeSeconds(comment.Timestamp / 1000).ToLocalTime();
             StringWriter messageWriter = new();
-            HttpUtility.HtmlDecode(comment.GetCommentMessage(), messageWriter);
+            HttpUtility.HtmlDecode(comment.GetCommentMessage(Database), messageWriter);
 
             string decodedMessage = messageWriter.ToString();
             string? url = Url.RouteUrl(ViewContext.RouteData.Values);