--- a/ProjectLighthouse.Servers.Website/Pages/Partials/SlotCardPartial.cshtml
+++ b/ProjectLighthouse.Servers.Website/Pages/Partials/SlotCardPartial.cshtml
@@ -7,11 +7,10 @@
 @using LBPUnion.ProjectLighthouse.Types.Users
 @using Microsoft.EntityFrameworkCore
 @model LBPUnion.ProjectLighthouse.Types.Entities.Level.SlotEntity
+@inject DatabaseContext Database
 
 @{
     UserEntity? user = (UserEntity?)ViewData["User"];
-
-    await using DatabaseContext database = DatabaseContext.CreateNewInstance();
 
     string slotName = HttpUtility.HtmlDecode(string.IsNullOrEmpty(Model!.Name) ? "Unnamed Level" : Model.Name);
 
@@ -26,8 +25,8 @@
 
     if (user != null)
     {
-        isQueued = await database.QueuedLevels.FirstOrDefaultAsync(h => h.SlotId == Model.SlotId && h.UserId == user.UserId) != null;
-        isHearted = await database.HeartedLevels.FirstOrDefaultAsync(h => h.SlotId == Model.SlotId && h.UserId == user.UserId) != null;
+        isQueued = await Database.QueuedLevels.FirstOrDefaultAsync(h => h.SlotId == Model.SlotId && h.UserId == user.UserId) != null;
+        isHearted = await Database.HeartedLevels.FirstOrDefaultAsync(h => h.SlotId == Model.SlotId && h.UserId == user.UserId) != null;
     }
 
     string callbackUrl = (string)ViewData["CallbackUrl"]!;
@@ -81,15 +80,15 @@
         }
 
         <div class="cardStatsUnderTitle">
-            <i class="pink heart icon" title="Hearts"></i> <span>@Model.Hearts</span>
+            <i class="pink heart icon" title="Hearts"></i> <span>@Model.Hearts(Database)</span>
             <i class="blue play icon" title="Plays"></i> <span>@Model.PlaysUnique</span>
-            <i class="green thumbs up icon" title="Yays"></i> <span>@Model.Thumbsup</span>
-            <i class="red thumbs down icon" title="Boos"></i> <span>@Model.Thumbsdown</span>
+            <i class="green thumbs up icon" title="Yays"></i> <span>@Model.Thumbsup(Database)</span>
+            <i class="red thumbs down icon" title="Boos"></i> <span>@Model.Thumbsdown(Database)</span>
 
             @if (Model.GameVersion == GameVersion.LittleBigPlanet1)
             {
                 <i class="yellow star icon" title="Star Rating"></i>
-                <span>@(Math.Round(Model.RatingLBP1 * 10) / 10)</span>
+                <span>@(Math.Round(Model.RatingLBP1(Database) * 10) / 10)</span>
             }
         </div>
         @if (Model.Creator != null)