--- a/ProjectLighthouse/Administration/Maintenance/RepeatingTasks/CleanupRoomsTask.cs
+++ b/ProjectLighthouse/Administration/Maintenance/RepeatingTasks/CleanupRoomsTask.cs
@@ -11,5 +11,5 @@ public class CleanupRoomsTask : IRepeatingTask
     public string Name => "Cleanup Rooms";
     public TimeSpan RepeatInterval => TimeSpan.FromSeconds(10);
     public DateTime LastRan { get; set; }
-    public Task Run(DatabaseContext database) => RoomHelper.CleanupRooms();
+    public Task Run(DatabaseContext database) => RoomHelper.CleanupRooms(database);
 }