Changes:
* Website pages now use `@inject` to get dependencies instead of manually instantiating them
* Program entry point now uses top-level statements
* Removed the obscure and undocumented startup command system
* Fixed ICommands to use an IServiceProvider instead of making their own DbContext
* Fixed cases of exception control flow in command logic
* Optimized some queries on the website to reduce database round trips
* Made StartupTasks async
* Fix a bad query in GameUserSlot that potentially caused way more rows than intended to be returned.
<details>
<summary>Benchmarks</summary>

## Before
```
Server Software:        Kestrel
Server Hostname:        localhost
Server Port:            10060

Document Path:          /
Document Length:        20228 bytes

Concurrency Level:      100
Time taken for tests:   375.067 seconds
Complete requests:      25000
Failed requests:        0
Total transferred:      509000000 bytes
HTML transferred:       505700000 bytes
Requests per second:    66.65 [#/sec] (mean)
Time per request:       1500.269 [ms] (mean)
Time per request:       15.003 [ms] (mean, across all concurrent requests)
Transfer rate:          1325.28 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       2
Processing:    77 1495 2992.2    549   63696
Waiting:       77 1494 2992.2    548   63696
Total:         78 1495 2992.2    550   63696

Percentage of the requests served within a certain time (ms)
  50%    550
  66%    899
  75%   1300
  80%   1638
  90%   3101
  95%   5906
  98%  12181
  99%  16600
 100%  63696 (longest request)
```

## After
```
Server Software:        Kestrel
Server Hostname:        localhost
Server Port:            10060

Document Path:          /
Document Length:        20230 bytes

Concurrency Level:      100
Time taken for tests:   190.686 seconds
Complete requests:      25000
Failed requests:        0
Total transferred:      509050000 bytes
HTML transferred:       505750000 bytes
Requests per second:    131.11 [#/sec] (mean)
Time per request:       762.743 [ms] (mean)
Time per request:       7.627 [ms] (mean, across all concurrent requests)
Transfer rate:          2607.01 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       3
Processing:    33  762  79.9    762    1207
Waiting:       30  759  79.8    759    1207
Total:         33  762  79.9    762    1207

Percentage of the requests served within a certain time (ms)
  50%    762
  66%    792
  75%    813
  80%    826
  90%    862
  95%    893
  98%    932
  99%    958
 100%   1207 (longest request)

```
</details>